+++ webots_ros2/CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog for package webots_ros2
 * Added component remapping parameter to WebotsController to rename PROTO components.
 * Added animation_{start,stop}_recording services to Ros2Supervisor node.
 * Added /Ros2Supervisor namespace to Ros2Supervisor node.
+* Fixed Python plugin termination on SIGINT call or simulation ends.
 
 2023.1.0 (2023-06-29)
 ------------------